requirements.txt
pytorch==1.5.0
numpy==1.21.2
scipy==1.4.1
sklearn==0.23.2
pandas
matplotlib
argparse
pickle
typing
os
sys
math
time
logging
# benchmarks
tensorflow=1.14.0
GPy=1.9.9

